rename to InterfaceSignal
